+++ internal/controller/daemonset_reconciler_test.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/go-logr/logr"
 	"github.com/thurgauerkb/cascader/internal/kinds"
-	"github.com/thurgauerkb/cascader/internal/testutils"
+	"github.com/thurgauerkb/cascader/test/testutils"
 
 	"github.com/stretchr/testify/assert"
 	appsv1 "k8s.io/api/apps/v1"
